requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
async-timeout==3.0.0
asyncpg==0.18.3
attrs==18.2.0
beautifulsoup4==4.6.3
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
discord.py==0.16.12
docutils==0.14
feedgenerator==1.9
idna==2.7
idna-ssl==1.1.0
ipaddress==1.0.22
Jinja2==2.10
logzero==1.5.0
lyricwikia==0.1.9
mail-parser==3.5.1
Markdown==2.6.11
MarkupSafe==1.0
multidict==4.4.0
mutagen==1.41.1
pafy==0.5.4
pathlib==1.0.1
pelican==3.7.1
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.20.0
simplejson==3.16.0
six==1.11.0
spotdl==1.0.0
spotipy==2.4.4
titlecase==0.12.0
unicode-slugify==0.1.3
Unidecode==1.0.22
urllib3==1.24.1
websockets==3.4
yarl==1.2.6
youtube-dl==2018.11.3
